+++ libft/stats/ft_get_high.c
@@ -6,7 +6,7 @@
 /*   By: amartino <amartino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/25 12:01:11 by amartino          #+#    #+#             */
-/*   Updated: 2019/11/25 13:08:27 by amartino         ###   ########.fr       */
+/*   Updated: 2020/03/12 14:41:47 by amartino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_get_high(int32_t *tab, size_t size)
 {
-	int32_t		high;
+	size_t		high;
 	size_t		i;
 
 	i = 0;
