+++ libft/stats/ft_get_low.c
@@ -6,7 +6,7 @@
 /*   By: amartino <amartino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/25 12:01:11 by amartino          #+#    #+#             */
-/*   Updated: 2020/01/17 15:33:32 by fkante           ###   ########.fr       */
+/*   Updated: 2020/03/12 14:41:37 by amartino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_get_low(int32_t *tab, size_t size)
 {
-	int32_t		low;
+	size_t		low;
 	size_t		i;
 
 	i = 0;
