+++ libft/stats/ft_get_low_in_range.c
@@ -6,7 +6,7 @@
 /*   By: fkante <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/17 15:15:01 by fkante            #+#    #+#             */
-/*   Updated: 2020/01/17 16:01:02 by fkante           ###   ########.fr       */
+/*   Updated: 2020/03/12 13:37:38 by amartino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_get_low_in_range(int32_t *tab, size_t start, size_t end)
 {
-	int32_t		low;
+	size_t		low;
 
 	low = start;
 	while (start < end)
